+++ src/victory-primitives/area.js
@@ -1,7 +1,9 @@
+/*eslint no-magic-numbers: ["error", { "ignore": [-1, 0, 1, 2] }]*/
 import React from "react";
 import PropTypes from "prop-types";
 import Collection from "../victory-util/collection";
 import Helpers from "../victory-util/helpers";
+import { defined, getXAccessor, getYAccessor, getY0Accessor, getAngleAccessor } from "./helpers";
 import { assign } from "lodash";
 import * as d3Shape from "d3-shape";
 import CommonProps from "./common-props";
@@ -18,58 +20,73 @@ export default class Area extends React.Component {
   };
 
   componentWillMount() {
-    const { style, areaPaths, linePaths } = this.calculateAttributes(this.props);
+    const { style, areaPath, linePath } = this.calculateAttributes(this.props);
     this.style = style;
-    this.areaPaths = areaPaths;
-    this.linePaths = linePaths;
+    this.areaPath = areaPath;
+    this.linePath = linePath;
   }
 
   shouldComponentUpdate(nextProps) {
-    const { style, areaPaths, linePaths } = this.calculateAttributes(nextProps);
+    const { style, areaPath, linePath } = this.calculateAttributes(nextProps);
     const { className, interpolation } = this.props;
     if (!Collection.allSetsEqual([
       [className, nextProps.className],
       [interpolation, nextProps.interpolation],
-      [linePaths, this.linePaths],
-      [areaPaths, this.areaPaths],
+      [linePath, this.linePath],
+      [areaPath, this.areaPath],
       [style, this.style]
     ])) {
       this.style = style;
-      this.areaPaths = areaPaths;
-      this.linePaths = linePaths;
+      this.areaPath = areaPath;
+      this.linePath = linePath;
       return true;
     }
     return false;
   }
 
-  calculateAttributes(props) {
-    const { style, data, active, scale } = props;
-    const dataSegments = this.getDataSegments(data);
-    const xScale = scale.x;
-    const yScale = scale.y;
+  getLineFunction(props) {
+    const { polar, scale } = props;
     const interpolation = this.toNewName(props.interpolation);
-    const areaFunction = d3Shape.area()
-      .curve(d3Shape[interpolation])
-      .x((d) => xScale(d._x1 !== undefined ? d._x1 : d._x))
-      .y1((d) => yScale(d._y1 !== undefined ? d._y1 : d._y))
-      .y0((d) => yScale(d._y0));
-    const lineFunction = d3Shape.line()
-      .curve(d3Shape[interpolation])
-      .x((d) => xScale(d._x1 !== undefined ? d._x1 : d._x))
-      .y((d) => yScale(d._y1));
-    return {
-      style: Helpers.evaluateStyle(assign({ fill: "black" }, style), data, active),
-      areaPaths: dataSegments.map((segment) => areaFunction(segment)),
-      linePaths: dataSegments.map((segment) => lineFunction(segment))
-    };
+    return polar ?
+      d3Shape.lineRadial()
+        .defined(defined)
+        .curve(d3Shape[`${interpolation}Closed`])
+        .angle(getAngleAccessor(scale))
+        .radius(getY0Accessor(scale)) :
+      d3Shape.line()
+        .defined(defined)
+        .curve(d3Shape[interpolation])
+        .x(getXAccessor(scale))
+        .y(getYAccessor(scale));
   }
 
-  getDataSegments(data) {
-    return Collection.splitArray(data, (datum) => {
-      const yDatum = datum.y1 !== undefined ? datum._y1 : datum._y;
+  getAreaFunction(props) {
+    const { polar, scale } = props;
+    const interpolation = this.toNewName(props.interpolation);
+    return polar ?
+      d3Shape.radialArea()
+        .defined(defined)
+        .curve(d3Shape[`${interpolation}Closed`])
+        .angle(getAngleAccessor(scale))
+        .outerRadius(getYAccessor(scale))
+        .innerRadius(getY0Accessor(scale)) :
+      d3Shape.area()
+        .defined(defined)
+        .curve(d3Shape[interpolation])
+        .x(getXAccessor(scale))
+        .y1(getYAccessor(scale))
+        .y0(getY0Accessor(scale));
+  }
 
-      return yDatum === null || typeof yDatum === "undefined";
-    }).filter((segment) => segment.length > 1);
+  calculateAttributes(props) {
+    const { style, data, active } = props;
+    const areaFunction = this.getAreaFunction(props);
+    const lineFunction = this.getLineFunction(props);
+    return {
+      style: Helpers.evaluateStyle(assign({ fill: "black" }, style), data, active),
+      areaPath: areaFunction(data),
+      linePath: lineFunction(data)
+    };
   }
 
   toNewName(interpolation) {
@@ -79,52 +96,52 @@ export default class Area extends React.Component {
   }
 
   // Overridden in victory-core-native
-  renderArea(paths, style, events) {
+  renderArea(path, style, events) {
     const areaStroke = style.stroke ? "none" : style.fill;
     const areaStyle = assign({}, style, { stroke: areaStroke });
-    const { role, shapeRendering, className } = this.props;
-    return paths.map((path, index) => {
-      return (
-        <path
-          key={`area-${index}`}
-          style={areaStyle}
-          shapeRendering={shapeRendering || "auto"}
-          role={role || "presentation"}
-          d={path}
-          className={className}
-          {...events}
-        />
-      );
-    });
+    const { role, shapeRendering, className, polar, origin } = this.props;
+    const transform = polar && origin ? `translate(${origin.x}, ${origin.y})` : undefined;
+    return (
+      <path
+        key={"area"}
+        style={areaStyle}
+        shapeRendering={shapeRendering || "auto"}
+        role={role || "presentation"}
+        d={path}
+        transform={transform}
+        className={className}
+        {...events}
+      />
+    );
   }
 
   // Overridden in victory-core-native
-  renderLine(paths, style, events) {
+  renderLine(path, style, events) {
     if (!style.stroke || style.stroke === "none" || style.stroke === "transparent") {
       return [];
     }
-    const { role, shapeRendering, className } = this.props;
+    const { role, shapeRendering, className, polar, origin } = this.props;
+    const transform = polar && origin ? `translate(${origin.x}, ${origin.y})` : undefined;
     const lineStyle = assign({}, style, { fill: "none" });
-    return paths.map((path, index) => {
-      return (
-        <path
-          key={`area-stroke-${index}`}
-          style={lineStyle}
-          shapeRendering={shapeRendering || "auto"}
-          role={role || "presentation"}
-          d={path}
-          className={className}
-          {...events}
-        />
-      );
-    });
+    return (
+      <path
+        key={"area-stroke"}
+        style={lineStyle}
+        shapeRendering={shapeRendering || "auto"}
+        role={role || "presentation"}
+        d={path}
+        transform={transform}
+        className={className}
+        {...events}
+      />
+    );
   }
 
   render() {
     const { events, groupComponent } = this.props;
-    const areas = this.renderArea(this.areaPaths, this.style, events);
-    const lines = this.renderLine(this.linePaths, this.style, events);
-    const children = [...lines, ...areas];
+    const area = this.renderArea(this.areaPath, this.style, events);
+    const line = this.renderLine(this.linePath, this.style, events);
+    const children = [line, area];
     return children.length === 1 ? children[0] : React.cloneElement(groupComponent, {}, children);
   }
 }